Module moonlander_gp::genetic

Genetic operation functions

Structs

ScoreCard
Immutable tagged list of scores.

SimpleFitness
Simple fitness result that only consists of a ScoreCard.

Weights
Parameters to the evolve function.

Traits

Fitness
Trait that models fitness for an individual

Functions

crossover_tree
Cross two trees.

evolve
Evolve an old generation into a new generation.

mutate_tree
Mutate a random node.

tournament_selection
Return the winner from a tournament of size N, randomly picked from the scored population.

Type Definitions

Scores
List of tuples of a score label and a score value.