Function moonlander_gp::genetic::evolve

pub fn evolve<P, F, S, R: Rng>(pop: Population<P, F>, weights: &Weights, rng: &mut R, selector: S) -> Population<P, F> where P: AstNode + Clone + Mutatable + Sync, F: Fitness + Send, S: for<'a> Fn(&'a Population<P, F>, &mut Rng) -> &'a P

Evolve an old generation into a new generation.

Uses the parameters in the Weights structure and the selection algorithm to pick and evolve individuals from the given population into a new one.