Struct moonlander_gp::genetic::ScoreCard

pub struct ScoreCard(_, _);

Immutable tagged list of scores.

The final score of an individual can be composed of many labeled subscores (for manual evaluation later on). Pass the scoresas a list of tuples with a static string and a score. The final score is the total of all subscores.

Example:

let score = ScoreCard::new(vec![
    ("food", 10.0),
    ("thirst", -5.0)
]);

Methods

impl ScoreCard

fn new(scores: Scores) -> ScoreCard

fn add(self, scores: Scores) -> ScoreCard

fn scores(&self) -> &Scores

fn total_score(&self) -> Number

fn is_empty(&self) -> bool